New stack merge

+++ compiler/thrift.go
@@ -20,23 +20,24 @@ const (
 	structLikeUnion     structLike = "union"
 )
 
-var thriftTypes = map[string]bool{
-	"bool":   true,
-	"byte":   true,
-	"i16":    true,
-	"i32":    true,
-	"i64":    true,
-	"double": true,
-	"string": true,
-	"binary": true,
-}
-
 func generateThriftIDL(dir string, frugal *parser.Frugal) (string, error) {
 	file := filepath.Join(dir, fmt.Sprintf("%s.thrift", frugal.Name))
+	if exists(file) {
+		// Trying to generate an intermediate Thrift IDL but the .thrift file
+		// already exists. First check if we generated it. If so, skip it. If
+		// not, return an error.
+		for _, intermediate := range globals.IntermediateIDL {
+			if intermediate == file {
+				return file, nil
+			}
+		}
+		return "", fmt.Errorf("Couldn't generate intermediate Thrift, file already exists: %s", file)
+	}
 	f, err := os.Create(file)
 	if err != nil {
 		return "", err
 	}
+	globals.IntermediateIDL = append(globals.IntermediateIDL, file)
 	defer f.Close()
 
 	contents := ""
@@ -48,7 +49,7 @@ func generateThriftIDL(dir string, frugal *parser.Frugal) (string, error) {
 		return "", err
 	}
 	contents += includes
-	contents += generateConstants(thrift)
+	contents += generateConstants(frugal)
 	contents += generateTypedefs(thrift.Typedefs)
 	contents += generateEnums(thrift.Enums)
 	contents += generateStructLikes(thrift.Structs, structLikeStruct)
@@ -71,36 +72,44 @@ func generateNamespaces(namespaces []*parser.Namespace) string {
 
 func generateIncludes(frugal *parser.Frugal) (string, error) {
 	contents := ""
+	// Recurse on includes
 	for _, incl := range frugal.Thrift.Includes {
 		include := incl.Value
-		if strings.HasSuffix(strings.ToLower(include), ".frugal") {
-			// Recurse on frugal includes
-			parsed, err := compile(filepath.Join(frugal.Dir, include))
-			if err != nil {
-				return "", err
-			}
-			// Lop off .frugal
-			includeBase := include[:len(include)-7]
-			frugal.ParsedIncludes[includeBase] = parsed
-
-			// Replace .frugal with .thrift
-			include = includeBase + ".thrift"
+		if !strings.HasSuffix(include, ".thrift") && !strings.HasSuffix(include, ".frugal") {
+			return "", fmt.Errorf("Bad include name: %s", include)
 		}
+
+		parsed, err := compile(filepath.Join(frugal.Dir, include),
+			strings.HasSuffix(include, ".thrift"), globals.Recurse)
+		if err != nil {
+			return "", err
+		}
+
+		// Lop off extension (.frugal or .thrift)
+		includeBase := include[:len(include)-7]
+
+		// Lop off path
+		includeName := filepath.Base(includeBase)
+
+		frugal.ParsedIncludes[includeName] = parsed
+
+		// Replace .frugal with .thrift
+		include = includeBase + ".thrift"
 		contents += fmt.Sprintf("include \"%s\"\n", include)
 	}
 	contents += "\n"
 	return contents, nil
 }
 
-func generateConstants(thrift *parser.Thrift) string {
+func generateConstants(frugal *parser.Frugal) string {
 	contents := ""
-	complexConstants := make([]*parser.Constant, 0, len(thrift.Constants))
+	complexConstants := make([]*parser.Constant, 0, len(frugal.Thrift.Constants))
 
-	for _, constant := range thrift.Constants {
+	for _, constant := range frugal.Thrift.Constants {
 		value := constant.Value
-		typeName := thrift.UnderlyingType(constant.Type.Name)
-		if isThriftPrimitive(typeName) {
-			if typeName == "string" {
+		underlyingType := frugal.UnderlyingType(constant.Type)
+		if parser.IsThriftPrimitive(underlyingType) {
+			if underlyingType.Name == "string" {
 				value = fmt.Sprintf(`"%s"`, value)
 			}
 		} else {
@@ -337,25 +346,14 @@ func generateServices(services []*parser.Service) string {
 	return contents
 }
 
-func generateThrift(frugal *parser.Frugal, idlDir, out, gen string, dryRun bool) error {
-	// Generate intermediate Thrift IDL.
-	idlFile, err := generateThriftIDL(idlDir, frugal)
-	if err != nil {
-		return err
-	}
-	globals.IntermediateIDL = append(globals.IntermediateIDL, idlFile)
-
-	if dryRun {
-		return nil
-	}
-
+func generateThrift(frugal *parser.Frugal, idlDir, file, out, gen string) error {
 	// Generate Thrift code.
 	args := []string{}
 	if out != "" {
 		args = append(args, "-out", out)
 	}
 	args = append(args, "-gen", gen)
-	args = append(args, idlFile)
+	args = append(args, file)
 	// TODO: make thrift command configurable
 	if out, err := exec.Command("thrift", args...).CombinedOutput(); err != nil {
 		fmt.Println(string(out))
@@ -386,9 +384,4 @@ func (e enumValues) Swap(i, j int) {
 
 func (e enumValues) Less(i, j int) bool {
 	return e[i].Value < e[j].Value
-}
-
-func isThriftPrimitive(typeName string) bool {
-	_, ok := thriftTypes[typeName]
-	return ok
 }
